+++ app/src/main/java/woowacourse/shopping/ui/cart/CartPresenter.kt
@@ -1,6 +1,7 @@
 package woowacourse.shopping.ui.cart
 
 import woowacourse.shopping.repository.CartRepository
+import woowacourse.shopping.ui.cart.uistate.CartUIState
 
 class CartPresenter(
     private val view: CartContract.View,
