+++ app/src/main/java/woowacourse/shopping/ui/productdetail/ProductDetailContract.kt
@@ -1,5 +1,7 @@
 package woowacourse.shopping.ui.productdetail
 
+import woowacourse.shopping.ui.productdetail.uistate.ProductDetailUIState
+
 interface ProductDetailContract {
     interface Presenter {
         fun loadProduct(productId: Long)
