+++ app/src/main/java/woowacourse/shopping/ui/productdetail/uistate/ProductDetailUIState.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.productdetail
+package woowacourse.shopping.ui.productdetail.uistate
 
 import woowacourse.shopping.domain.Product
