+++ app/src/main/res/layout/item_product.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="product"
-            type="woowacourse.shopping.ui.products.ProductUIState" />
+            type="woowacourse.shopping.ui.products.uistate.ProductUIState" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
